Dust monitor, pressure control system
distributed system over CAN bus inside a mining vehicle to measure the dust level, control exhaust fans and measure the differential pressure to make equilibrium. The system was using HMI display that allows the user to set thresholds and check valves states. It's optionally allows the user to do firmware update for all and different component over CAN bus using a firmware on the USB on one of these boards. It's optionally connected to a smartphone app to send logs of the system