multi-protocol-device.png
this project that uses WiFi, M2M NBIoT BG95 LTE module to connect to the internet and get protobuf data from the server then show on eInk display. It support the connection over LoRaWAN in areas that don't have LTE or WiFi networks. It was ultra low power device that is dedicated to send/receive notification to users.